separed model files
