+++ app/controllers/access_controller.py
@@ -3,7 +3,7 @@
 from fastapi import HTTPException, status, Depends
 from sqlalchemy.orm import Session
 from app.services.access_service import authenticate_user, create_access_token, get_password_hash
-from app.models import Usuario
+from app.models.user_db_models import Usuario
 from app.database import get_db
 
 def login_user(username: str, clave: str, db: Session = Depends(get_db)):
