+++ app/database/seeds.py
@@ -1,6 +1,7 @@
 from datetime import date
 from app.database.database import SessionLocal, Base, engine
-from app.database.models import Persona, Direccion, Telefono, Producto, Poliza, Pago, Siniestro
+from app.database.user_db_models import Persona, Direccion, Telefono
+from app.database.insurance_db_models import Producto, Poliza, Pago, Siniestro
 
 Base.metadata.create_all(bind=engine)
 db = SessionLocal()
