+++ app/models/__init__.py
@@ -1,1 +1,3 @@
-from .models import *
+from .email_db_models import *
+from .user_db_models import *
+from .insurance_db_models import *
